Move component TokyoAlertCard from 'components' to 'components/index/SiteTopUpper'
