+++ components/index/SiteTopUpper.vue
@@ -31,8 +31,8 @@ import { MetaInfo } from 'vue-meta'
 
 import MonitoringCommentCard from '@/components/index/SiteTopUpper/MonitoringCommentCard.vue'
 import StaticInfoCard from '@/components/index/SiteTopUpper/StaticInfoCard.vue'
+import TokyoAlertCard from '@/components/index/SiteTopUpper/TokyoAlertCard.vue'
 import PageHeader from '@/components/PageHeader.vue'
-import TokyoAlertCard from '@/components/TokyoAlertCard.vue'
 import WhatsNew from '@/components/WhatsNew.vue'
 import Data from '@/data/data.json'
 import News from '@/data/news.json'
